Fixed errors and converted the engine from dynamically linked to static

+++ LostSouls/src/Platform/Windows/WindowsInput.cpp
@@ -31,7 +31,7 @@ namespace LostSouls {
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
 
-		return std::pair<float, float>(xpos, ypos);
+		return std::pair<float, float>((float)xpos, (float)ypos);
 	}
 
 	float WindowsInput::GetMouseXImpl()
